+++ src/App.jsx
@@ -1,7 +1,9 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import Footer from './components/layout/Footer';
 import Home from './pages/Home';
 import Nosotros from './pages/Nosotros';
 import Cotizar from './pages/Cotizar';
+import Ingenieria from './pages/soluciones/Ingenieria';
 import WhatsAppButton from './components/ui/WhatsAppButton';
 
 function App() {
@@ -15,6 +17,7 @@ function App() {
                     <Route path="/" element={<Home />} />
                     <Route path="/nosotros" element={<Nosotros />} />
                     <Route path="/cotizar" element={<Cotizar />} />
+                    <Route path="/soluciones/ingenieria" element={<Ingenieria />} />
                 </Routes>
                 <WhatsAppButton />
             </div>